Добавил AppBar для экрана с аккаунтами и выделил AccountCard в отдельный файл
